### 24.04.02 ~ 24.04.08
### 24.04.02

- [x]  리뷰 목록 동적으로 불러오기 및 화면 개발

### 24.04.03

- [x]  관리자 상품관리 초기 페이지 개발
- 상품 목록 동적으로 가져오기
- [x]  상품등록 폼 페이지 개발

### 24.04.04 ~ 04.05

- [x]  관리자 상품등록 기능 개발
- 에디터 api를 받아 쓰는데 생각보다 파일 전송이 쉽지 않음.

### 24.04.06

- [x]  상품등록 페이지 누락 입력값 추가 (제조사, 등장작품, 상품종류)
- [x]  상품 수정기능
- [x]  상품 삭제기능

### 24.04.07~04.08

- [x]  상품목록 동적 페이지 개발
- [x]  상품목록 페이징 처리
- [x]  상품목록 검색기능
- [x]  상품관리 상품목록 페이징 처리
- [x]  상품관리 상품목록 검색기능

+++ src/main/java/com/anitomo/utilities/FileUploadHandler.java
@@ -5,6 +5,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 @Slf4j
@@ -43,15 +44,48 @@ public class FileUploadHandler
         return "success";
     }
 
-    public void deleteFile(String code, String fileName, String boardType)
+    public void deleteFile(String code, String[] fileNameArr, String boardType)
     {
         if(boardType.equals("review"))
-            path = "C:\\development\\projects\\anitomo\\images\\reviews\\"+code+"\\"+fileName;
+        {
+            for (String fileName : fileNameArr)
+            {
+                path = "C:\\development\\projects\\anitomo\\images\\reviews\\"+code+"\\"+fileName;
+                file = new File(path);
+                file.delete();
+            }
+        }
+        else
+        {
+            path = "C:\\development\\projects\\anitomo\\images\\items\\"+code+"\\";
+            file = new File(path);
+            File[] fileArr = file.listFiles();
+            List<String> fileNameList = Arrays.asList(fileNameArr);
+
+            for (File tmpFile : fileArr)
+                if(!fileNameList.contains(tmpFile.getName()))
+                    tmpFile.delete();
+        }
+    }
+
+    public void deleteDirectory(String code, String boardType)
+    {
+        if(boardType.equals("review"))
+            path = "C:\\development\\projects\\anitomo\\images\\reviews\\"+ code +"\\";
 
         file = new File(path);
 
-        log.info(path);
+        if(file.isDirectory())
+        {
+            File[] fileArr = file.listFiles();
 
-        file.delete();
+            // 폴더가 비어있지 않으면 삭제가 안 되므로 파일 삭제 수행
+            if(fileArr != null)
+                for (File deleteTarget : fileArr)
+                    deleteTarget.delete();
+
+            // 폴더 삭제
+            file.delete();
+        }
     }
 }
